+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Remove recorder overlay minimize/restore controls and hotkey handling.
+
 ## [1.5.9] - 2026-05-20
 
 ### Fixed